+++ freemarker-generator-cli/src/test/java/org/apache/freemarker/generator/cli/ManualTest.java
@@ -25,8 +25,12 @@ public class ManualTest extends AbstractMainTest {
 
     // private static final String CMD = "-V";
     private static final String CMD =
-            "-t src/app/examples/templates/nginx/confluence/nginx-config-parser.ftl -s src/app/examples/data/nginx";
+            // "-i 'Hello${tools.system.systemProperty(\"NAME\")}' -DNAME=sgoeschl"
+            // "-i ${tools.yaml.toYaml(json)} -m json=src/app/examples/data/json/swagger-spec.json"
+            "-i '${tools.yaml.toYaml(json)}' -m json=https://petstore.swagger.io/v2/swagger.json"
+            // "-t src/app/examples/templates/nginx/confluence/nginx-config-parser.ftl -s src/app/examples/data/nginx";
             // "-t src/app/examples/templates/datasources.ftl";
+            ;
 
     @Override
     public String execute(String commandLine) throws IOException {